Organized files. Read 'Organization Format' for further details.
