+++ WePool/angular.json
@@ -24,9 +24,7 @@
               "src/assets"
             ],
             "styles": [
-
-              "./node_modules/@angular/material/prebuilt-themes/deeppurple-amber.css",
-
+              "./node_modules/@angular/material/prebuilt-themes/indigo-pink.css",
               "src/styles.css"
             ],
             "scripts": []
@@ -91,9 +89,7 @@
               "src/assets"
             ],
             "styles": [
-
-              "./node_modules/@angular/material/prebuilt-themes/deeppurple-amber.css",
-
+              "./node_modules/@angular/material/prebuilt-themes/indigo-pink.css",
               "src/styles.css"
             ],
             "scripts": []
