+++ WePool/package.json
@@ -23,7 +23,7 @@
     "@angular/router": "~9.0.3",
     "@googlemaps/js-api-loader": "^1.13.2",
     "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
+    "tslib": "^1.14.1",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
@@ -31,9 +31,7 @@
     "@angular/cli": "~9.0.4",
     "@angular/compiler-cli": "~9.0.3",
     "@angular/language-service": "~9.0.3",
-
     "@types/google.maps": "^3.47.3",
-
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
