+++ WePool/src/app/account-files/index.ts
@@ -1,0 +1,1 @@
+export { LoginComponent } from "./login/login.component";
