+++ WePool/src/app/app.module.ts
@@ -2,29 +2,18 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
+import { AppComponent } from './app-files/app.component';
 
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MatToolbarModule } from '@angular/material/toolbar';
-import { MatButtonModule } from '@angular/material/button';
-import { MatIconModule } from '@angular/material/icon';
-import { MatGridListModule } from '@angular/material/grid-list';
-import { MatCardModule } from '@angular/material/card';
-import { CarpoolsComponent } from './carpools/carpools.component';
-import { HomeComponent } from './home/home.component';
-import { ProfileComponent } from './profile/profile.component';
-import { AdministrationComponent } from './administration/administration.component';
-import { ModerationComponent } from './moderation/moderation.component';
-import { LoginComponent } from './login/login.component';
-import { MapPOCComponent } from './map-poc/map-poc.component';
-import { MatSliderModule } from '@angular/material/slider';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MatInputModule } from '@angular/material/input';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { HttpClientModule } from '@angular/common/http';
-import { LocationAutocompleteComponent } from './location-autocomplete/location-autocomplete.component';
+import { ProfileComponent } from './profile/profile.component'; //Will change when previous pull request comes in
 import { ToolbarComponent } from './toolbar/toolbar.component';
 
+import { HttpClientModule } from '@angular/common/http';
+import { MapPOCComponent, LocationAutocompleteComponent, CarpoolsComponent } from './carpool-files';
+import { AdministrationComponent, ModerationComponent } from './moderation-files';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { LoginComponent } from './account-files';
+import { MaterialModule } from './modules/material.module';
+import { HomeComponent } from './home-files';
 
 @NgModule({
   declarations: [
@@ -38,25 +27,13 @@ import { ToolbarComponent } from './toolbar/toolbar.component';
     MapPOCComponent,
     LocationAutocompleteComponent,
     ToolbarComponent
-
   ],
   imports: [
-    MatToolbarModule,
-    MatButtonModule,
-    MatIconModule,
-    MatGridListModule,
-    MatCardModule,
+    BrowserAnimationsModule,
     BrowserModule,
     AppRoutingModule,
-    BrowserAnimationsModule,
-    MatSliderModule,
-    BrowserAnimationsModule,
-    FormsModule,
-    ReactiveFormsModule,
-    MatInputModule,
-    MatFormFieldModule,
     HttpClientModule,
-
+    MaterialModule
   ],
   providers: [],
   bootstrap: [AppComponent]
