+++ WePool/src/app/carpool-files/index.ts
@@ -1,0 +1,3 @@
+export { CarpoolsComponent } from './carpools/carpools.component';
+export { LocationAutocompleteComponent } from './location-autocomplete/location-autocomplete.component';
+export { MapPOCComponent } from './map-poc/map-poc.component';
