+++ WePool/src/app/home-files/home/home.component.css
