+++ WePool/src/app/home-files/index.ts
@@ -1,0 +1,1 @@
+export { HomeComponent } from "./home/home.component";
