+++ WePool/src/app/moderation-files/index.ts
@@ -1,0 +1,2 @@
+export { AdministrationComponent } from "./administration/administration.component";
+export { ModerationComponent } from "./moderation/moderation.component";
