+++ WePool/src/app/moderation-files/moderation/moderation.component.ts
