+++ WePool/src/app/modules/material.module.ts
@@ -1,0 +1,30 @@
+import { NgModule } from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatSliderModule } from '@angular/material/slider';
+import { MatGridListModule } from '@angular/material/grid-list';
+import { MatToolbarModule } from '@angular/material/toolbar';
+
+@NgModule({
+  exports: [
+    MatButtonModule,
+    MatCardModule,
+    MatIconModule,
+    MatInputModule,
+    MatSliderModule,
+    MatGridListModule,
+    MatToolbarModule
+  ],
+  imports: [
+    MatButtonModule,
+    MatCardModule,
+    MatIconModule,
+    MatInputModule,
+    MatSliderModule,
+    MatGridListModule,
+    MatToolbarModule
+  ]
+})
+export class MaterialModule { }
